requirements.txt
# HealthCheck Project Requirements
# Install with: pip install -r requirements.txt

# System monitoring
psutil>=5.9.0

# HTTP requests for Telegram
requests>=2.31.0

# Google Gemini AI SDK
google-genai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0

# Modern GUI framework (replacing Tkinter)
flet>=0.21.0

# Build tool for Windows executable
pyinstaller>=6.0.0
